+++ src/main/java/aries/Aries.java
@@ -1,5 +1,7 @@
 package aries;
 
+import aries.command.Command;
+import aries.command.CommandParser;
 import aries.storage.Storage;
 import aries.task.TaskList;
