+++ src/main/java/aries/command/ByeCommand.java
@@ -1,5 +1,6 @@
-package aries;
+package aries.command;
 
+import aries.Ui;
 import aries.task.TaskList;
 
 public class ByeCommand implements Command {
