+++ src/main/java/aries/command/Command.java
@@ -1,5 +1,7 @@
-package aries;
+package aries.command;
 
+import aries.AriesException;
+import aries.Ui;
 import aries.task.TaskList;
 
 public interface Command {
