+++ src/main/java/aries/command/DeleteCommand.java
@@ -1,22 +1,24 @@
-package aries;
+package aries.command;
 
+import aries.AriesException;
+import aries.Ui;
 import aries.task.Task;
 import aries.task.TaskList;
 import aries.util.IndexHandling;
 
-public class MarkCommand implements Command {
+public class DeleteCommand implements Command {
     private String index;
 
-    public MarkCommand(String index) {
+    public DeleteCommand(String index) {
         this.index = index;
     }
 
     @Override
     public boolean execute(TaskList tasks, Ui ui) throws AriesException {
         int index = IndexHandling.getValidIndex(this.index, tasks.size());
-        Task taskToMark = tasks.get(index);
-        taskToMark.markAsDone();
-        ui.marked(taskToMark, true);
+        Task taskToDelete = tasks.get(index);
+        tasks.remove(index);
+        ui.deleted(taskToDelete, tasks);
         return true;
     }
 }
