+++ src/main/java/aries/command/MarkCommand.java
@@ -1,22 +1,24 @@
-package aries;
+package aries.command;
 
+import aries.AriesException;
+import aries.Ui;
 import aries.task.Task;
 import aries.task.TaskList;
 import aries.util.IndexHandling;
 
-public class DeleteCommand implements Command {
+public class MarkCommand implements Command {
     private String index;
 
-    public DeleteCommand(String index) {
+    public MarkCommand(String index) {
         this.index = index;
     }
 
     @Override
     public boolean execute(TaskList tasks, Ui ui) throws AriesException {
         int index = IndexHandling.getValidIndex(this.index, tasks.size());
-        Task taskToDelete = tasks.get(index);
-        tasks.remove(index);
-        ui.deleted(taskToDelete, tasks);
+        Task taskToMark = tasks.get(index);
+        taskToMark.markAsDone();
+        ui.marked(taskToMark, true);
         return true;
     }
 }
